Refactor CLI subcommands into separate files
